User_Interface/requirements.txt
Streamlit == 1.35.0
plotly == 5.22.0
psycopg2-binary == 2.9.9
geopy == 2.4.1